+++ packages/aws-amplify-react/src/API/GraphQL/Connect.js
@@ -1,6 +1,6 @@
 import regeneratorRuntime from 'regenerator-runtime/runtime';
 import React, { Component } from 'react';
-import { API } from '../../Categories';
+import API from '@aws-amplify/api';