+++ packages/core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.1-unstable.1"></a>
+## [1.0.1-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1...@aws-amplify/core@1.0.1-unstable.1) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.1-unstable.0"></a>
+## [1.0.1-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1...@aws-amplify/core@1.0.1-unstable.0) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
 <a name="1.0.1-next.2"></a>
 ## [1.0.1-next.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1-next.1...@aws-amplify/core@1.0.1-next.2) (2018-06-28)